+++ linters/bandit/bandit_annotations.js
@@ -1,7 +1,7 @@
 // Copyright 2018 VMware, Inc.
 // SPDX-License-Identifier: BSD-2-Clause
 
-const { getAnnotationLevel } = require('../annotations_levels')
+const { getAnnotationLevel } = require('../../annotations_levels')
 
 /**
  * Create bandit Annotations
