+++ linters/gosec/gosec.js
@@ -1,19 +1,20 @@
-// Copyright 2019 VMware, Inc.
+// Copyright 2018 VMware, Inc.
 // SPDX-License-Identifier: BSD-2-Clause
 
-const cache = require('../cache')
-const report = require('../tslint/tslint_report')
+const cache = require('../../cache')
 
-module.exports = class TSLint {
+const report = require('./gosec_report')
+
+module.exports = class Gosec {
   get name () {
-    return 'tslint'
+    return 'gosec'
   }
 
   /**
    * The name of the generated report file
    */
   get reportFile () {
-    return 'tslint_output.json'
+    return 'gosec_output.json'
   }
 
   get defaultReport () {
@@ -26,7 +27,7 @@ module.exports = class TSLint {
    * @returns {string[]} Filtered list of file names
    */
   filter (files) {
-    return files.filter(name => name.endsWith('.js') || name.endsWith('.ts'))
+    return files.filter(name => name.endsWith('.go'))
   }
 
   /**
@@ -35,7 +36,7 @@ module.exports = class TSLint {
    * @param {string} prID PR id in repository
    */
   workingDirectoryForPR (repoID, prID) {
-    return cache.getBranchPath(repoID, prID, 'tslint')
+    return cache.getBranchPath(repoID, prID, 'gosec')
   }
 
   /**
@@ -44,7 +45,7 @@ module.exports = class TSLint {
    * @param {string} reportPath Path to the report file relative to working directory
    */
   args (files, reportPath) {
-    return ['-c', '../../../tslint/tslint.json', '--format', 'json', '-o', reportPath, ...files]
+    return ['-fmt=json', '-out', reportPath, './...']
   }
 
   /**
@@ -52,7 +53,29 @@ module.exports = class TSLint {
    * @param {Buffer} data The raw linter results data
    */
   parseResults (data) {
-    return JSON.parse(data)
+    let parsedData = JSON.parse(data)
+    let syntaxErrors = parsedData['Golang errors']
+
+    if (syntaxErrors) {
+      let filePaths = Object.keys(syntaxErrors)
+
+      for (let path of filePaths) {
+        for (let error of syntaxErrors[path]) {
+          let errAnnotation = {
+            severity: 'HIGH',
+            confidence: 'HIGH',
+            rule_id: 'ERROR',
+            details: 'Syntax error',
+            file: path,
+            code: `${error.error}`,
+            line: error.line
+          }
+          parsedData.Issues.push(errAnnotation)
+        }
+      }
+    }
+
+    return parsedData
   }
 
   /**
