+++ linters/index.js
@@ -1,9 +1,9 @@
 // Copyright 2018 VMware, Inc.
 // SPDX-License-Identifier: BSD-2-Clause
 
-const Bandit = require('./bandit')
-const Gosec = require('./gosec')
-const TSLint = require('./tslint')
+const Bandit = require('./bandit/bandit')
+const Gosec = require('./gosec/gosec')
+const TSLint = require('./tslint/tslint')
 
 module.exports = {
   BANDIT: new Bandit(),
